+++ provision.sh
@@ -5,11 +5,10 @@
 # ===================================================================
 
 
-# -------------------------------------------------------------------
-# Parametrization (for all environment)
-# -------------------------------------------------------------------
-export BOOT2DOCKER_EXTENSION_DIR="/var/lib/boot2docker/extension"
-export BOOT2DOCKER_EXTENSION_LIB_DIR="$BOOT2DOCKER_EXTENSION_DIR/lib"
+# ------------------------------------------------------------
+# Environment definition (params and so on)
+# ------------------------------------------------------------
+source /var/lib/boot2docker/extension/environment.sh
 
 
 # -------------------------------------------------------------------